move locations from Attendee info to info tab
